+++ src/Win32Application/MessageBoxLogger.cpp
@@ -13,7 +13,7 @@ namespace Mengine
 	{
 	}
 	//////////////////////////////////////////////////////////////////////////
-	void MessageBoxLogger::log( EMessageLevel _level, uint32_t _flag, const Char * _data, size_t _count )
+	void MessageBoxLogger::log( EMessageLevel _level, uint32_t _flag, const Char * _data, uint32_t _count )
 	{	
 		(void)_level;
         (void)_flag;
